requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Vector Databases
chromadb==0.4.22
pinecone-client==3.0.2
faiss-cpu==1.7.4
annoy==1.17.3

# Embeddings & LLM
openai==1.12.0
anthropic==0.18.0
google-generativeai==0.3.2
langchain==0.1.6
langchain-community==0.0.20
sentence-transformers==2.3.1

# Data Processing
numpy==1.26.3
pandas==2.1.4
tiktoken==0.5.2

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1

# API & Web
httpx==0.26.0
aiofiles==23.2.1
python-multipart==0.0.6
websockets==12.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
locust==2.20.1

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0